+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 -   Adds a suggester to autocomplete paths for the save and open file inputs.
 -   Adds a validator to validate paths for the save and open file inputs.
 -   `textual-textarea` now requires `textual` >=0.27.0
+-   Adds reactive properties to the textarea for `selection_anchor` position and 
+    `selected_text`.
 
 ## [0.3.3] - 2023-07-28